Refactor libs for cleaner linkage
