--- source/bython/lib.cpp
@@ -1,1 +1,0 @@
-// #include "lib.hpp"
