--- source/bython/llvm_backend/CMakeLists.txt
@@ -1,8 +1,0 @@
-target_sources(bython_lib PRIVATE
-    builtin.cpp
-    intrinsic.cpp
-    llvm.cpp
-    symbols.cpp
-    type_system.cpp
-)
-#add_subdirectory(codegen)
