--- source/compiler.cpp
@@ -1,10 +1,0 @@
-#include <iostream>
-#include <memory>
-#include <string>
-
-#include <llvm/Support/CommandLine.h>
-
-auto main(int argc, char* argv[]) -> int
-{
-  namespace cl = llvm::cl;
-}
